TRANSLATOR and VALIDATOR feature added. Test:
(let* ((model #~0)
       (view #λ(let ((value ~model))
                 (when-commit ()
                   (format t "VIEW: update based on model: ~A~%~%" value))))
       (view-feedback #λ(when-let (event (feedback-event-of model))
                          (format t "VIEW: correct your input and try again: ~A ~S~%" ~model event))))
  (declare (ignore view view-feedback))

  (setf (validator-of model)
        λ(unless (evenp ~model)
           (format t "(VALIDATOR-OF MODEL): ~A is not an even number; aborting!~%" ~model)
           (pulse (feedback-event-of model) :not-even)
           (throw :abort nil)))

  (setf (translator-of model)
        (lambda (input)
          (typecase input
            (number input)
            (string (parse-integer input))
            (t (error "Don't know what to do with ~S~%" input)))))

  (loop :for i :from 1 :upto 11
     :do (catch :abort
           (with-sync ()
             (setf ~model (format nil "~D" i)))))

  (format t "~%always an even number: ~A~%" ~model))


VIEW: update based on model: 0

(VALIDATOR-OF MODEL): 1 is not an even number; aborting!
VIEW: correct your input and try again: 1 :NOT-EVEN
VIEW: update based on model: 2

(VALIDATOR-OF MODEL): 3 is not an even number; aborting!
VIEW: correct your input and try again: 3 :NOT-EVEN
VIEW: update based on model: 4

(VALIDATOR-OF MODEL): 5 is not an even number; aborting!
VIEW: correct your input and try again: 5 :NOT-EVEN
VIEW: update based on model: 6

(VALIDATOR-OF MODEL): 7 is not an even number; aborting!
VIEW: correct your input and try again: 7 :NOT-EVEN
VIEW: update based on model: 8

(VALIDATOR-OF MODEL): 9 is not an even number; aborting!
VIEW: correct your input and try again: 9 :NOT-EVEN
VIEW: update based on model: 10

(VALIDATOR-OF MODEL): 11 is not an even number; aborting!
VIEW: correct your input and try again: 11 :NOT-EVEN

always an even number: 10